Refactoring of the tests
