+++ tests/java/unit/com/dataiku/dss/formats/kml/sample_kmz.kmz
